Extract MonthlyCheckoutMessage component
